+++ PrimitiveModifier.java
@@ -1,5 +1,5 @@
 public interface PrimitiveModifier extends JSONSerializable{
-	void calculate(int timeStep);
+	void calculate(double timeStep);
 	void setPrimitive(ScenePrimitive sp);
 	PrimitiveModifier clone(ScenePrimitive sp);
 }
